refactor: extract common functions
